deployment/requirements.txt
seaborn
pandas
matplotlib
plotly
Pillow
scikit-learn==1.2.2
streamlit_option_menu